# Conflicts:
#	src/main/java/de/uniks/beastopia/teaml/controller/ingame/IngameController.java
#	src/main/java/de/uniks/beastopia/teaml/modules/HttpModule.java
#	src/main/java/de/uniks/beastopia/teaml/service/DataCache.java
#	src/main/resources/de/uniks/beastopia/teaml/assets/lang_de.properties
#	src/main/resources/de/uniks/beastopia/teaml/assets/lang_en.properties

+++ src/main/java/de/uniks/beastopia/teaml/rest/Trainer.java
@@ -14,7 +14,7 @@ public record Trainer(
         List<String> team,
         List<Integer> encounteredMonsterTypes,
         List<String> visitedAreas,
-        int coins,
+        float coins,
         String area,
         int x,
         int y,
